requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
boto3==1.37.37
botocore==1.37.37
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap5==2025.4
cssselect2==0.8.0
dj-database-url==2.3.0
Django==4.2.20
django-bootstrap-v5==1.0.11
django-crispy-forms==2.4
django-environ==0.12.0
django-formtools==2.5.1
django-jazzmin==3.0.1
django-storages==1.14.6
fontawesomefree==6.6.0
fonttools==4.57.0
gunicorn==23.0.0
html5lib==1.1
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
pip3-autoremove==1.2.2
psycopg2-binary==2.9.10
pycairo==1.27.0
pycparser==2.22
pydyf==0.6.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.11.5
setuptools==75.5.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
urllib3==2.4.0
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
